Remove refine and collapse
