+++ langchain/src/chains/combine_documents/index.ts
@@ -1,2 +1,1 @@
 export { createStuffDocumentsChain } from "./stuff.js";
-export { createRefineDocumentsChain } from "./refine.js";
