+++ langchain/src/chains/combine_documents/reduce.ts
@@ -1,5 +1,4 @@
 import { Document } from "../../document.js";
-import { collapseDocsMetadata } from "./collapse.js";
 
 /**
  * Splits a list of documents into sublists based on a maximum token limit.
@@ -57,4 +56,23 @@ export async function collapseDocs(
 ): Promise<Document> {
   const result = await combineDocumentFunc(docs);
   return { pageContent: result, metadata: collapseDocsMetadata(docs) };
+}
+
+function collapseDocsMetadata(docs: Document[]): Document["metadata"] {
+  const combinedMetadata: Record<string, string> = {};
+  for (const key in docs[0].metadata) {
+    if (key in docs[0].metadata) {
+      combinedMetadata[key] = String(docs[0].metadata[key]);
+    }
+  }
+  for (const doc of docs.slice(1)) {
+    for (const key in doc.metadata) {
+      if (key in combinedMetadata) {
+        combinedMetadata[key] += `, ${doc.metadata[key]}`;
+      } else {
+        combinedMetadata[key] = String(doc.metadata[key]);
+      }
+    }
+  }
+  return combinedMetadata;
 }
